requirements.txt
tensorflow==1.11.0
scipy==1.1.0
numpy==1.15.4
Pillow==3.4.0
Flask==1.0.2
Werkzeug==0.14.1
cassandra_driver==3.15.1
